change `modulo` to `remainder` and add `modulo`
The function formerly known as `modulo` was actually `remainder`, so
rename it, but add `modulo` to the library. This change is not
technically less backward-compatible than most changes, since it
removes `modulo` from `zuo/kernel`.

Thanks to @soegaard for noticing the problem and providing a first cut
at a repair.