Safe-for-space with linked enviorments variant of Zuo
Plain Zuo doesn't prune environments when forming closures, which
means that it has the space complexity of S_tail in Clinger (ICFP'98).
This experimental variant of Zuo has a space complexity that is at
least as good as S_sfs, and also at least as good as linked closures
(section 13 in Clinger).

Instead of the conventional implementaiton of space safety by pruning
an environment whenever a closure is formed, the Zuo implementation
gets space safety through the GC, which prunes an environment when
travsering it in combination with a live-variable set (i.e., the free
variables of a closure), where the same environment can be paired with
multiple such sets, and sharing of environment structure is preserved.
Environments are binary trees instead of linked lists, so they also
provide nearly constant-time access to variables in the environment.